Implement blackmarket and a feature to refresh it

+++ src/main/java/zip/sodium/marketplace/data/listing/Listing.java
@@ -4,4 +4,4 @@ import org.bukkit.OfflinePlayer;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
-public record Listing(@NotNull OfflinePlayer seller, @NotNull ItemStack stack, int price) { }
+public record Listing(@NotNull OfflinePlayer seller, @NotNull ItemStack stack, double price) { }
